moved to extension_resolver.rs to config extensins mod
