+++ crates/goose-server/src/routes/agent.rs
@@ -14,6 +14,7 @@ use goose::agents::ExtensionLoadResult;
 
 use base64::Engine;
 use goose::agents::ExtensionConfig;
+use goose::config::resolve_extensions_for_new_session;
 use goose::config::{Config, GooseMode};
 use goose::model::ModelConfig;
 use goose::prompt_template::render_global_file;
@@ -22,7 +23,7 @@ use goose::recipe::Recipe;
 use goose::recipe_deeplink;
 use goose::session::extension_data::ExtensionState;
 use goose::session::session_manager::SessionType;
-use goose::session::{resolve_extensions_for_new_session, EnabledExtensionsState, Session};
+use goose::session::{EnabledExtensionsState, Session};
 use goose::{
     agents::{extension::ToolInfo, extension_manager::get_parameter_names},
     config::permission::PermissionLevel,
